move readers functionality to examples
